+++ event-sourcing/event-repository/event-repository-jdbc/src/test/java/uk/gov/justice/services/eventsourcing/repository/jdbc/eventlog/EventLogConverterTest.java
@@ -1,10 +1,14 @@
 package uk.gov.justice.services.eventsourcing.repository.jdbc.eventlog;
 
+import static java.time.temporal.ChronoUnit.SECONDS;
+import static org.exparity.hamcrest.date.ZonedDateTimeMatchers.within;
 import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static uk.gov.justice.services.messaging.JsonObjectMetadata.metadataFrom;
 
 import uk.gov.justice.services.common.converter.StringToJsonObjectConverter;
+import uk.gov.justice.services.common.util.DateTimeProvider;
 import uk.gov.justice.services.eventsourcing.common.exception.InvalidStreamIdException;
 import uk.gov.justice.services.messaging.DefaultJsonEnvelope;
 import uk.gov.justice.services.messaging.JsonEnvelope;
@@ -45,6 +49,7 @@ public class EventLogConverterTest {
         eventLogConverter = new EventLogConverter();
         eventLogConverter.stringToJsonObjectConverter = new StringToJsonObjectConverter();
         eventLogConverter.jsonObjectEnvelopeConverter = new JsonObjectEnvelopeConverter();
+        eventLogConverter.dateTimeProvider = new DateTimeProvider();
     }
 
     @Test
@@ -59,6 +64,7 @@ public class EventLogConverterTest {
         assertThat(eventLog.getSequenceId(), equalTo(SEQUENCE_ID));
         JSONAssert.assertEquals(METADATA_JSON, eventLog.getMetadata(), false);
         JSONAssert.assertEquals(expectedPayloadAsJsonString, eventLog.getPayload(), false);
+        assertThat(eventLog.getDateCreated(), is(within(5L, SECONDS, new DateTimeProvider().now())));
     }
 
     @Test(expected = InvalidStreamIdException.class)
@@ -77,7 +83,7 @@ public class EventLogConverterTest {
     }
 
     private EventLog createEventLog() {
-        return new EventLog(ID, STREAM_ID, SEQUENCE_ID, NAME, METADATA_JSON, PAYLOAD_JSON);
+        return new EventLog(ID, STREAM_ID, SEQUENCE_ID, NAME, METADATA_JSON, PAYLOAD_JSON, new DateTimeProvider().now());
     }
 
     private JsonEnvelope createTestEnvelope() throws IOException {